+++ modules/core/src/main/java/org/locationtech/jts/algorithm/hull/PolygonConcaveHull.java
@@ -16,7 +16,6 @@ import java.util.PriorityQueue;
 import org.locationtech.jts.algorithm.Orientation;
 import org.locationtech.jts.geom.Coordinate;
 import org.locationtech.jts.geom.CoordinateArrays;
-import org.locationtech.jts.geom.CoordinateList;
 import org.locationtech.jts.geom.Envelope;
 import org.locationtech.jts.geom.Geometry;
 import org.locationtech.jts.geom.GeometryFactory;
@@ -43,7 +42,7 @@ public class PolygonConcaveHull {
    * the vertices forming the angle are in CW orientation.
    */
   private final Coordinate[] vertex;
-  private VertexRing vertexRing;
+  private LinkedRing vertexRing;
   
   /**
    * Indexing vertices improves corner intersection testing performance.
@@ -70,15 +69,15 @@ public class PolygonConcaveHull {
     return toGeometry();
   }
   
-  private void init(Coordinate[] vertex) { 
+  private void init(Coordinate[] ring) { 
     //-- ensure ring is CW
-    if (Orientation.isCCW(vertex)) {
-      vertex = vertex.clone();
-      CoordinateArrays.reverse(vertex);
+    if (Orientation.isCCW(ring)) {
+      ring = ring.clone();
+      CoordinateArrays.reverse(ring);
     }
     
-    vertexRing = new VertexRing(vertex);
-    vertexIndex = new VertexSequencePackedRtree(vertex);
+    vertexRing = new LinkedRing(ring);
+    vertexIndex = new VertexSequencePackedRtree(ring);
     
     cornerQueue = new PriorityQueue<Corner>();
     for (int i = 0; i < vertexRing.size(); i++) {
@@ -106,11 +105,11 @@ public class PolygonConcaveHull {
         && vertexRing.size() > 3) {
       Corner corner = cornerQueue.poll();
       //-- a corner may no longer be valid due to removal of adjacent corners
-      if (vertexRing.isRemoved(corner))
+      if (corner.isRemoved(vertexRing))
         continue;
       //System.out.println(corner.toLineString(vertexList));
       /**
-       * Concave/Flat corner - remove it if safe.
+       * Corner is concave or flat - remove it if possible.
        */
       if ( isRemovable(corner) ) {
         removeCorner(corner, cornerQueue);
@@ -167,105 +166,6 @@ public class PolygonConcaveHull {
     return fact.createPolygon(fact.createLinearRing(coords));
   }
 
-  private static class VertexRing {
-    
-    private static final int NO_VERTEX_INDEX = -1;
-
-    private final Coordinate[] vertex;
-    private int[] next = null;
-    private int[] prev = null;
-    private int size;
-    
-    public VertexRing(Coordinate[] pts) {
-      vertex = pts;
-      size = pts.length - 1;
-      next = createNextLinks(size);
-      prev = createPrevLinks(size);
-    }
-
-    public int size() {
-      return size;
-    }
-
-    public Coordinate getCoordinate(int index) {
-      return vertex[index];
-    }
-
-    private static int[] createNextLinks(int size) {
-      int[] next = new int[size];
-      for (int i = 0; i < size; i++) {
-        next[i] = i + 1;
-      }
-      next[size - 1] = 0;
-      return next;
-    }
-    
-    private static int[] createPrevLinks(int size) {
-      int[] prev = new int[size];
-      for (int i = 0; i < size; i++) {
-        prev[i] = i - 1;
-      }
-      prev[0] = size - 1;
-      return prev;
-    }
-    
-    public boolean isConvex(int index) {
-      Coordinate pp = vertex[prev[index]];
-      Coordinate p = vertex[index];
-      Coordinate pn = vertex[next[index]];
-      return Orientation.CLOCKWISE == Orientation.index(pp, p, pn);
-    }
-
-    public int next(int i) {
-      return next[i];
-    }
-
-    public int prev(int i) {
-      return prev[i];
-    }
-
-    public boolean hasVertex(int index) {
-      return index < prev.length 
-          && prev[index] != NO_VERTEX_INDEX;
-    }
-    public double area(int index) {
-      Coordinate pp = vertex[prev[index]];
-      Coordinate p = vertex[index];
-      Coordinate pn = vertex[next[index]];
-      return Triangle.area(pp,  p,  pn);
-    }
-    
-    public void remove(int index) {
-      int iprev = prev[index];
-      int inext = next[index];
-      next[iprev] = inext;
-      prev[inext] = iprev;
-      prev[index] = NO_VERTEX_INDEX;
-      next[index] = NO_VERTEX_INDEX;
-      size--;
-    }
-
-    public boolean isRemoved(Corner corner) {
-      return isCornerRemoved(corner.getIndex(), corner.getPrev(), corner.getNext());
-    }
-    
-    public boolean isCornerRemoved(int index, int iprev, int inext) {
-      return prev[index] != iprev
-          || next[index] != inext;
-    }
-    
-    public Coordinate[] getCoordinates() {
-      CoordinateList coords = new CoordinateList();
-      for (int i = 0; i < vertex.length - 1; i++) {
-        if (prev[i] != NO_VERTEX_INDEX) {
-          coords.add(vertex[i].copy(), false);
-        }
-      }
-      coords.closeRing();
-      return coords.toCoordinateArray();
-    }
-  }
-  
   private static class Corner implements Comparable<Corner> {
     private int index;
     private int prev;
@@ -302,26 +202,30 @@ public class PolygonConcaveHull {
       return Double.compare(area, o.area);
     }
     
-    public Envelope envelope(VertexRing vertexList) {
-      Coordinate pp = vertexList.getCoordinate(prev);
-      Coordinate p = vertexList.getCoordinate(index);
-      Coordinate pn = vertexList.getCoordinate(next);
+    public Envelope envelope(LinkedRing ring) {
+      Coordinate pp = ring.getCoordinate(prev);
+      Coordinate p = ring.getCoordinate(index);
+      Coordinate pn = ring.getCoordinate(next);
       Envelope env = new Envelope(pp, pn);
       env.expandToInclude(p);
       return env;
     }
     
-    public boolean intersects(Coordinate v, VertexRing vertexList) {
-      Coordinate pp = vertexList.getCoordinate(prev);
-      Coordinate p = vertexList.getCoordinate(index);
-      Coordinate pn = vertexList.getCoordinate(next);
+    public boolean intersects(Coordinate v, LinkedRing ring) {
+      Coordinate pp = ring.getCoordinate(prev);
+      Coordinate p = ring.getCoordinate(index);
+      Coordinate pn = ring.getCoordinate(next);
       return Triangle.intersects(pp, p, pn, v);
     }
     
-    public Geometry toLineString(VertexRing vertexList) {
-      Coordinate pp = vertexList.getCoordinate(prev);
-      Coordinate p = vertexList.getCoordinate(index);
-      Coordinate pn = vertexList.getCoordinate(next);
+    public boolean isRemoved(LinkedRing ring) {
+      return ring.prev(index) != prev || ring.next(index) != next;
+    }
+    
+    public Geometry toLineString(LinkedRing ring) {
+      Coordinate pp = ring.getCoordinate(prev);
+      Coordinate p = ring.getCoordinate(index);
+      Coordinate pn = ring.getCoordinate(next);
       return (new GeometryFactory()).createLineString(
           new Coordinate[] { safeCoord(pp), safeCoord(p), safeCoord(pn) });
     }
