+++ modules/core/src/test/java/org/locationtech/jts/algorithm/hull/PolygonConcaveHullTest.java
@@ -1,3 +1,14 @@
+/*
+ * Copyright (c) 2021 Martin Davis.
+ *
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License 2.0
+ * and Eclipse Distribution License v. 1.0 which accompanies this distribution.
+ * The Eclipse Public License is available at http://www.eclipse.org/legal/epl-v20.html
+ * and the Eclipse Distribution License is available at
+ *
+ * http://www.eclipse.org/org/documents/edl-v10.php.
+ */
 package org.locationtech.jts.algorithm.hull;
 
 import org.locationtech.jts.geom.Geometry;
@@ -26,6 +37,11 @@ public class PolygonConcaveHullTest extends GeometryTestCase {
   public void testFlat() {
     checkHull("POLYGON ((10 10, 10 90, 90 90, 90 50, 90 10, 50 10, 10 10))", 
         0.7, "POLYGON ((10 10, 10 90, 90 90, 90 10, 10 10))");
+  }
+
+  public void testInnerHull() {
+    checkHull("POLYGON ((11 14, 2 31, 18 29, 25 17, 38 16, 29 5, 19 11, 11 0, 0 10, 11 14))", 
+        0.5, "POLYGON ((10 10, 10 90, 90 90, 90 10, 10 10))");
   }
 
   private void checkHull(String wkt, double vertexCountFraction, String wktExpected) {
